add Patient class and modified Person
